atualizacao

+++ db/migrate/20161114194832_create_os_deb_tecnicos.rb
@@ -1,10 +1,10 @@
 class CreateOsDebTecnicos < ActiveRecord::Migration
   def change
     create_table :os_deb_tecnicos do |t|
-      t.decimal :debt_id
-      t.decimal :os_id
+      t.integer :debt_id
+      t.integer :os_id
 
       t.timestamps null: false
     end
-  end
+end
 end
