+++ db/migrate/20161114194846_create_tarefa_deb_tecnicos.rb
@@ -1,8 +1,8 @@
 class CreateTarefaDebTecnicos < ActiveRecord::Migration
   def change
     create_table :tarefa_deb_tecnicos do |t|
-      t.decimal :tare_id
-      t.decimal :debt_id
+      t.integer :tare_id
+      t.integer :debt_id
 
       t.timestamps null: false
     end
